requirements.txt
apache-libcloud==3.2.0
appdirs==1.4.4
asgiref==3.3.1
audioread==2.1.8
boto3==1.17.5
botocore==1.20.5
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colour==0.1.5
conllu==4.2.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
deepspeech==0.8.2
defusedxml==0.6.0
deprecation==2.1.0
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
django-filter==2.4.0
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
djangorestframework-xml==2.0.0
doccano==1.0.2
drf-yasg==1.20.0
ecdsa==0.14.1
environs==9.2.0
et-xmlfile==1.0.1
eyeD3==0.9.5
fastapi==0.63.0
ffmpeg==1.4
filelock==3.0.12
filetype==1.0.7
furl==2.1.0
future==0.18.2
google-api-core==1.22.4
google-auth==1.26.0
google-cloud==0.34.0
google-cloud-core==1.6.0
google-cloud-speech==2.0.0
google-cloud-translate==3.0.2
googleapis-common-protos==1.52.0
grpcio==1.32.0
h11==0.11.0
halo==0.0.30
hmmlearn==0.2.4
idna==2.10
importlib-metadata==3.4.0
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
kiwisolver==1.2.0
libcst==0.3.13
librosa==0.8.0
llvmlite==0.34.0
lml==0.1.0
log-symbols==0.0.14
MarkupSafe==1.1.1
marshmallow==3.9.1
matplotlib==3.3.2
mkl-service==2.3.0
mypy-extensions==0.4.3
numba==0.51.2
numpy==1.19.2
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.5
orderedmultidict==1.0.1
packaging==20.4
pandas==1.1.5
pdf2image==1.14.0
Pillow @ file:///tmp/build/80754af9/pillow_1594307298074/work
plotly==4.12.0
pocketsphinx==0.1.15
pooch==1.2.0
proto-plus==1.10.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pyAudioAnalysis==0.3.6
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==1.7.3
pydub==0.24.1
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-xlsx==0.6.0
PyJWT==1.7.1
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
Resemblyzer==0.1.1.dev0
retrying==1.3.3
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.3.2
selenium==3.141.0
sentencepiece==0.1.95
seqeval==1.2.2
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
SoundFile==0.10.3.post1
spectralcluster==0.1.0
spinners==0.0.24
sqlparse==0.4.1
starlette==0.13.6
termcolor==1.1.0
textract-trp==0.1.3
texttable==1.6.3
threadpoolctl==2.1.0
tokenizers==0.10.1
torch==1.6.0
torchvision==0.4.2
tqdm==4.51.0
transformers==4.3.2
typing==3.7.4.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.10
uvicorn==0.13.2
webrtcvad==2.0.10
whitenoise==5.2.0
zipp==3.4.0
